Moved create theme into bus logic
